Performs mutational enrichment analysis for a given clinical feature.

Performs paiwise and groupwise fisher exact tests to find differentially enriched genes for every factor within a clinical feature.

clinicalEnrichment(
  maf,
  clinicalFeature = NULL,
  annotationDat = NULL,
  minMut = 5,
  useCNV = TRUE
)

Arguments

maf
MAF object

clinicalFeature
columns names from `clinical.data` slot of MAF to be analysed for.

annotationDat
If MAF file was read without clinical data, provide a custom data.frame or a tsv file with a column containing Tumor_Sample_Barcodes along with clinical features. Default NULL.

minMut
Consider only genes with minimum this number of samples mutated. Default 5.

useCNV
whether to include copy number events. Only applicable when MAF is read along with copy number data. Default TRUE if available.

Value

result list containing p-values

See also

Examples

if (FALSE) {
laml.maf = system.file('extdata', 'tcga_laml.maf.gz', package = 'maftools')
laml.clin = system.file('extdata', 'tcga_laml_annot.tsv', package = 'maftools')
laml = read.maf(maf = laml.maf, clinicalData = laml.clin)
clinicalEnrichment(laml, 'FAB_classification')
}