Estimate number of signatures based on cophenetic correlation metric

Estimate number of signatures based on cophenetic correlation metric

estimateSignatures(
  mat,
  nMin = 2,
  nTry = 6,
  nrun = 10,
  parallel = 4,
  pConstant = NULL,
  verbose = TRUE,
  plotBestFitRes = FALSE
)

Arguments

mat
Input matrix of diemnsion nx96 generated by trinucleotideMatrix

nMin
Minimum number of signatures to try. Default 2.

nTry
Maximum number of signatures to try. Default 6.

nrun
numeric giving the number of run to perform for each value in range. Default 5

parallel
Default 4. Number of cores to use.

pConstant
A small positive value to add to the matrix. Use it ONLY if the functions throws an non-conformable arrays error

verbose
Default TRUE

plotBestFitRes
plots consensus heatmap for range of values tried. Default FALSE

Value

a list with NMF.rank object and summary stats.

Details

This function decomposes a non-negative matrix into n signatures. Extracted signatures are compared against 30 experimentally validated signatures by calculating cosine similarity. See http://cancer.sanger.ac.uk/cosmic/signatures for details.

See also

Examples

if (FALSE) {
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf)
laml.tnm <- trinucleotideMatrix(maf = laml, ref_genome = 'BSgenome.Hsapiens.UCSC.hg19', prefix = 'chr',
add = TRUE, useSyn = TRUE)
library("NMF")
laml.sign <- estimateSignatures(mat = laml.tnm, plotBestFitRes = FALSE, nMin = 2, nTry = 3, nrun = 2, pConstant = 0.01)
}