Plot gistic results along linearized chromosome

A genomic plot with segments highlighting signififcant Amplifications and Deletion regions.

gisticChromPlot(
  gistic = NULL,
  fdrCutOff = 0.1,
  markBands = NULL,
  color = NULL,
  ref.build = "hg19",
  cytobandOffset = 0.01,
  txtSize = 0.8,
  cytobandTxtSize = 0.6,
  maf = NULL,
  mutGenes = NULL,
  y_lims = NULL,
  mutGenesTxtSize = 0.6
)

Arguments

gistic
an object of class GISTIC generated by readGistic

fdrCutOff
fdr cutoff to use. Default 0.1

markBands
any cytobands to label. Default top 5 lowest q values.

color
colors for Amp and Del events.

ref.build
reference build. Could be hg18, hg19 or hg38.

cytobandOffset
if scores.gistic file is given use this to adjust cytoband size.

txtSize
label size for lables

cytobandTxtSize
label size for cytoband

maf
an optional maf object

mutGenes
mutated genes from maf object to be highlighted

y_lims
Deafult NULL. A vector upper and lower y-axis limits

mutGenesTxtSize
Default 0.6

Value

nothing

Examples

all.lesions <- system.file("extdata", "all_lesions.conf_99.txt", package = "maftools")
amp.genes <- system.file("extdata", "amp_genes.conf_99.txt", package = "maftools")
del.genes <- system.file("extdata", "del_genes.conf_99.txt", package = "maftools")
scores.gistic <- system.file("extdata", "scores.gistic", package = "maftools")
laml.gistic = readGistic(gisticAllLesionsFile = all.lesions, gisticAmpGenesFile = amp.genes, gisticDelGenesFile = del.genes, gisticScoresFile = scores.gistic)
gisticChromPlot(laml.gistic)