Plot gistic results.

takes output generated by readGistic and draws a plot similar to oncoplot.

gisticOncoPlot(
  gistic = NULL,
  top = NULL,
  bands = NULL,
  showTumorSampleBarcodes = FALSE,
  gene_mar = 5,
  barcode_mar = 6,
  sepwd_genes = 0.5,
  sepwd_samples = 0.25,
  clinicalData = NULL,
  clinicalFeatures = NULL,
  sortByAnnotation = FALSE,
  sampleOrder = NULL,
  annotationColor = NULL,
  bandsToIgnore = NULL,
  removeNonAltered = TRUE,
  colors = NULL,
  SampleNamefontSize = 0.6,
  fontSize = 0.8,
  legendFontSize = 1.2,
  annotationFontSize = 1.2,
  borderCol = "white",
  bgCol = "#CCCCCC"
)

Arguments

gistic
an GISTIC object generated by readGistic

top
how many top cytobands to be drawn. defaults to all.

bands
draw oncoplot for these bands. Default NULL.

showTumorSampleBarcodes
logical to include sample names.

gene_mar
Default 5

barcode_mar
Default 6

sepwd_genes
Default 0.5

sepwd_samples
Default 0.25

clinicalData
data.frame with columns containing Tumor_Sample_Barcodes and rest of columns with annotations.

clinicalFeatures
columns names from `clinicalData` to be drawn in the plot. Dafault NULL.

sortByAnnotation
logical sort oncomatrix (samples) by provided `clinicalFeatures`. Defaults to FALSE. column-sort

sampleOrder
Manually speify sample names for oncolplot ordering. Default NULL.

annotationColor
list of colors to use for clinicalFeatures. Default NULL.

bandsToIgnore
do not show these bands in the plot Default NULL.

removeNonAltered
Logical. If TRUE removes samples with no mutations in the oncoplot for better visualization. Default FALSE.

colors
named vector of colors Amp and Del events.

SampleNamefontSize
font size for sample names. Default 0.6

fontSize
font size for cytoband names. Default 0.8

legendFontSize
font size for legend. Default 1.2

annotationFontSize
font size for annotations. Default 1.2

borderCol
Default "white"

bgCol
Default "#CCCCCC"

Value

None.

Details

Takes gistic file as input and plots it as a matrix. Any desired annotations can be added at the bottom of the oncoplot by providing annotation

See also

Examples

all.lesions <- system.file("extdata", "all_lesions.conf_99.txt", package = "maftools")
amp.genes <- system.file("extdata", "amp_genes.conf_99.txt", package = "maftools")
del.genes <- system.file("extdata", "del_genes.conf_99.txt", package = "maftools")
scores.gistic <- system.file("extdata", "scores.gistic", package = "maftools")
laml.gistic = readGistic(gisticAllLesionsFile = all.lesions, gisticAmpGenesFile = amp.genes, gisticDelGenesFile = del.genes, gisticScoresFile = scores.gistic)
gisticOncoPlot(laml.gistic)