Plots results from oncodrive

Takes results from oncodrive and plots them as a scatter plot. Size of the gene shows number of clusters (hotspots), x-axis can either be an absolute number of variants accumulated in these clusters or a fraction of total variants found in these clusters. y-axis is fdr values transformed into -log10 for better representation. Labels indicate Gene name with number clusters observed.

plotOncodrive(
  res = NULL,
  fdrCutOff = 0.05,
  useFraction = FALSE,
  colCode = NULL,
  bubbleSize = 1,
  labelSize = 1
)

Arguments

res
results from oncodrive

fdrCutOff
fdr cutoff to call a gene as a driver.

useFraction
if TRUE uses a fraction of total variants as X-axis scale instead of absolute counts.

colCode
Colors to use for indicating significant and non-signififcant genes. Default NULL

bubbleSize
Size for bubbles. Default 2.

labelSize
font size for labelling genes. Default 1.

Value

Nothing

See also

Examples

laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf)
laml.sig <- oncodrive(maf = laml, AACol = 'Protein_Change', minMut = 5)
plotOncodrive(res = laml.sig, fdrCutOff = 0.1)