Plot Transition and Trasnversion ratios.

Takes results generated from titv and plots the Ti/Tv ratios and contributions of 6 mutational conversion classes in each sample.

plotTiTv(
  res = NULL,
  plotType = "both",
  sampleOrder = NULL,
  color = NULL,
  showBarcodes = FALSE,
  textSize = 0.8,
  baseFontSize = 1,
  axisTextSize = c(1, 1),
  plotNotch = FALSE
)

Arguments

res
results generated by titv

plotType
Can be 'bar', 'box' or 'both'. Defaults to 'both'

sampleOrder
Sample names in which the barplot should be ordered. Default NULL

color
named vector of colors for each coversion class.

showBarcodes
Whether to include sample names for barplot

textSize
fontsize if showBarcodes is TRUE. Deafult 2.

baseFontSize
font size. Deafult 1.

axisTextSize
text size x and y tick labels. Default c(1,1).

plotNotch
logical. Include notch in boxplot.

Value

None.

See also

Examples

laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf)
laml.titv = titv(maf = laml, useSyn = TRUE)
plotTiTv(laml.titv)