Performs sample stratification based on signature contribution and enrichment analysis.

Performs k-means clustering to assign signature to samples and performs enrichment analysis.

signatureEnrichment(maf, sig_res, minMut = 5, useCNV = FALSE, fn = NULL)

Arguments

maf
an MAF object used for signature analysis.

sig_res
Signature results from extractSignatures

minMut
Consider only genes with minimum this number of samples mutated. Default 5.

useCNV
whether to include copy number events. Only applicable when MAF is read along with copy number data. Default TRUE if available.

fn
basename for output file. Default NULL.

Value

result list containing p-values

See also